Added new Action to Declaratively setup GitHub labels
This action is a way to define the labels for a GitHub repo inside
the repo itself. It runs a fairly small node app to see the diff of the
.github/labels.json file and will only update the various labels it needs
to. As well, currently it makes the labels in a repo match the json file,
so it will delete labels that are on the repo but not in json.